Extract `process_order_or_checkout_with_transaction` to utils
